move auth middleware to seperate files
